requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.4
django-ckeditor==6.7.2
django-ckeditor-5==0.2.15
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-js-asset==3.0.1
django-quill-editor==0.1.42
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
sqlparse==0.5.3
tzdata==2024.2
